Always send -fexternal-interpreter when using iserv in ghci rules
Summary: When relying on the iserv binary, we don't always send `-fexternal-interpreter` from further up. This ensures that ghc will be instructed to use the iserv binary when we generate the template for it.

Test Plan: Run on a handful of ghci flavors and ghci rules. Saw previously breaking one pass.

fbshipit-source-id: 2f47cd5