Add swiftmodule information to swift object files
Summary:
lldb expects that the swift AST will be contained in a .swift_ast section
inside of swift object files. The preferred method of insertion
is to store the AST in a wrapper object file, which is then linked
into the final libraries.

Test Plan: com.facebook.buck.apple.AppleLibraryIntegrationTest

Reviewed By: milend

fbshipit-source-id: d9286f0