Make ThrowingSupplier.get() throw E
Summary:
ThrowingSupplier has an odd pattern different from
ThrowingConsumer/ThrowingFunction where it implements Supplier with a
method that wraps the thrown exception (and has a separate
throwingGet()).

The only place where ThrowingSupplier is currently used is a place that
already can throw the type of exception that its supplier can throw
(IOException).

This changes ThrowingSupplier to no longer implement Supplier and to
better follow the pattern of ThrowingConsumer/ThrowingFunction.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 482398f