Improve guidance for working with application binary interfaces (ABI) in Java
Summary: Incorporate guidance on how to design builds so that application binary interfaces (ABIs) are efficient.

Test Plan:
No code tests required; these are updates to the public-facing Buck documentation for application binary interfaces.

Before sending out my diff, I reviewed the Buck documentation in rendered form using a local web server. The script for doing this in in the Buck source tree. Here is the procedure.

Start the web server:

    $ cd ~/local/buck/docs
    $ ./soyweb-local.sh

    Listening on /0:0:0:0:0:0:0:0:9811

Go to your web browser and listen on the port where the server is running In the above example, that would be:

    http://localhost:9811/

I also run `aspell`, which is a command-line spell checker. On OSX, you can install `aspell` using `brew`:

    brew install aspell

Reviewed By: bobyangyf

fbshipit-source-id: 3b781d2