Buffer `buck audit rules` output, and flush stderr on superconsole before printing
Summary: Buffer the output of buck audit rules, and then print it all at the end of the command. This is due to how superconsoleeventlistener works (it stops printing new stderr lines after you dirty the stdout stream). We print skylark errors via the eventbus, so we need to make sure that all errors get printed before we dirty the stdout stream. This also adds a flush event so that we can ensure everything's written before printing to stdout.

Test Plan:
Ran buck audit rules on a small file that also prints out a line:
- WIthout this patch, error lines weren't printed, audit was
- Without the flush, but with the buffer: Sometimes got the error line, sometimes didn't
- With both: The errors are printed first, then the audit output

Also verified that buck build --show-output works.

Reviewed By: ttsugriy

fbshipit-source-id: eb8be98