versions: create and submit FJP actions atomically
Summary:
Previously, we'd add recursive FJP actions to a global map for others to grab
and subsequently submit them to the pool non-atomically.  This allowed other
workers to potentially attempt to `join()` tasks which hadn't been submitted
yet, which may lead to some blocking.

This diff makes it so we atomically create and submit tasks.

Test Plan: CI

Reviewed By: bobyangyf

fbshipit-source-id: a8e7318