versions: add a timeout and retries for version target graph building
Summary:
Until we can diagnose the deadlock issue occasionally hitting the FJP inside
the `VersionedTargetGraphBuilder` this wraps the process in a (generous) timeout
and retries it a few times while logging the occurence.

This should mean we stop affecting users while getting some numbers on how
common the deadlock is.

Test Plan: CI

Reviewed By: elsteveogrande

fbshipit-source-id: 86f2dbf