Sort slaves by number of active jobs.
Summary: This diff sorts the slaves by the number of active jobs, and hides the idle slaves from the console. Buck does not print idle or even downloading threads in the console for local build, and that is appreciated by devs. We should keep the behavior consistent.

Test Plan:
CI and -
eyes

Reviewed By: aiked

fbshipit-source-id: 3292dbf