Merge multiple serial event upload requests into one.
Summary:
Currently we lose slave side statistics for logging in more than half of stampede runs. Even when
the remote build finishes first, we lose this data one-third of the times. This is because we only
wait for a maximum of 5 seconds on the client side after the build has finished. This diff reduces
the time spent in closing off the SlaveEventBusListener from `> 6` seconds to `< 2` seconds. This
won't fix all our woes, but it's critical in getting a lot more data logged in our tables.

Test Plan: CI

Reviewed By: ruibm

fbshipit-source-id: a1c2f0f