Buck incremental project generation
Summary:
For focus4idea, we would like to incrementally refresh a single module or small set of modules in order to avoid regenerating the entire
project whenever a `BUCK` file changes. This diff includes a strawman implementation (that is fairly inefficient but straightforward).
In this implementation, we restrict the set of output files to just the ones matching the targets which were used to invoke `buck project`.

If we decide to move forward with this, the next steps are to limit the parsing and generation to just the top-level targets so that we actually
see a speedup.

Test Plan:
Running
`buck project //java/com/facebook/litho: --intellij-aggregation-mode none devEnv/focus/config/buckconfig --file-with-list-of-generated-files /tmp/regenerate.txt --update`
gives the following list of files written: P59053015 (vs ~40k files for `buck project fb4a`)

new unit tests

Reviewed By: styurin

fbshipit-source-id: 3daa404