Make caching aware of target graph only deps
Summary:
There are cases during `BuildRule` generation where we expect target graph "only" deps (great name) to be in the `BuildRuleResolver`.

Note that this is orthogonal to the main issues I'm working on fixing. This is something new that showed up this week in adhoc testing.

Test Plan: Added unit test.

Reviewed By: bobyangyf

fbshipit-source-id: 9dfdf96