Use ToolchainProvider from cell in BuildRuleCreationContext
Summary:
Right now `BuildRuleCreationContext` uses a `ToolchainProvider` from the root cell, but it should pass `ToolchainProvider` from the same cell as the target.

The solution is to pass `CellProvider` to `DefaultTargetNodeToBuildRuleTransformer` and use it to find the target's cell and get its `ToolchainProvider`.

Most of the changes is plumbing to pass `CellProvider` instead of `ToolchainProvider`.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 9708c24