Add BuildRuleStrategy hooks to build engine
Summary:
This adds a small hook in CachingBuildRuleBuilder to inject a
custom strategy for running the build steps of a buildrule. It
introduces a couple of interfaces for this:

BuildExecutor: This contains a single function `executeCommands(...)`.
This has the responsibility of basically calling `getBuildSteps()` and
running those steps.

BuildExecutorRunner: Provides methods to build a rule with the default
executor or with a custom executor. The ability to run the rule with the
default executor makes it easy for a custom strategy to run some rules
as normal and others in a custom way.

BuildRuleStrategy: This has one important method `build(...)`.
This is passed a BuildExecutorRunner and is expected to call one of its
two execute methods.

This change adds an `Optional<BuildRuleStrategy>` to the
CachingBuildEngine. This is created based on the new
ModernBuildRuleConfig (which supports just two cases NONE and
DEBUG_PASSTHROUGH which are both rather trivial).

Test Plan:
Added unit tests for CachingBuildEngine use of BuildRuleStrategy
Added integration tests for the ModernBuildRule strategies

Reviewed By: styurin

fbshipit-source-id: 3b25e8d