Add DEBUG_RECONSTRUCT strategy
Summary:
This is just a simple build strategy that takes a buildrule, serializes
it, deserializes, gets build steps from the deserialized version and
runs them.

Splits ModernBuildRule.getBuildSteps() into two parts:
1. ModernBuildRule.stepsForBuildable() - this allows to get the build
steps from a Buildable
2. ModernBuildRule.recordOutputs() - this records the outputs for a
ModernBuildRule.

The reconstructing strategy runs (1) on the deserialized version and (2)
on the original version. This matches how it would be done if the
actual step running were happening outside of the buck process (i.e. we
still need to record the outputs to the current process' build
engine).

Test Plan: Integration tests previously added cover this.

Reviewed By: ttsugriy

fbshipit-source-id: 6a7a684