Add methods to get the classpath for Buck
Summary:
In BundledExternalProcessLauncher, there was some logic for determining
the classpath required to run a Buck subprocess. This moves that into
BuckClasspath and improves it.

When running buck tests via `buck test`, the classpath will include the
classpath for the tests themselves (it's really difficult to write tests
of subprocess functionality otherwise).

Added BuckClasspath.getBootstrapClasspath(). When running within
Intellij, this will just return the full normal classpath, but when
running standalone buck, this will return the limited classpath that is
used for ClassLoaderBootstrapper.

Test Plan: Added tests.

Reviewed By: ttsugriy

fbshipit-source-id: 3ae3c49