Add InputsDigestBuilder
Summary:
For remote execution, the inputs are specified as a merkle tree with the
hash of the root specifying the entire inputs (all the file/directories
are added to a CAS and then materialized from it based on the root
digest).

This is a helper for constructing such an input digest (and a listing of
all the data that is required to be within the CAS for materializing
that input).

Test Plan: Added some tests.

Reviewed By: ttsugriy

fbshipit-source-id: 64f953c