TargetGraphHashing: use ListeningExecutor to calculate hashes
Summary:
Our current implementatino of target hashing, which uses a combination of a
guava `LoadingCache` and a fixed thead pool, will busy loop if a thread dies
while calculating the hash of a node, as upstream dependents continuously poll
until the (failed) dependency has finished successfully, which it never will.

One potential fix is to have the cache store an either type containing both the
failure and succces, and have the nodes check and propagate the error from their
dependencies.  This opts instead to just use a `ListenableExecutor` to avoid
the busy waiting in the loop completely (where it peels off an item, checks if
it's ready, and returns it to the queue if not) and propagate exceptions without
any special handling.

Test Plan: Added a unittest where hash calculation throws.

Reviewed By: bobyangyf

fbshipit-source-id: 4744f7e