Fix DelegatingClassloader behavior on Davlik
Summary:
DelegatingClassloader was calling `new DexFile(path/to/dex)` which uses the default opt location.
This is not writeable by our random app, so the fix is to produce a correct dexopt location.

Test Plan: `/scripts/exopackage_test_single.sh` used to fail on a Dalvik emulator (tested on KitKat). Now it passes

Reviewed By: styurin

fbshipit-source-id: 48b6957