Log stack traces from dry run failures
Summary:
Pass test failure information from dry runs when running with JUnitRunner.

Sometimes JUnitRunner fails to instantiate the test class, which will cause the dry run to return `<test_class>$initializationError` as a result.

The diff should be safe, because the status of dry run was already returned as failed, just not stack trace was attached.

Test Plan: Creating a test which throws InitializationError, verify that stacktrace of the error is attached on dry run with given test.

Reviewed By: bobyangyf, styurin

fbshipit-source-id: dd0aa5a