rust: skip Rust deps for a NativeLinkable Rust target
Summary:
When generating a Rust crate for linking with C++ - a cdylib type
crate - skip all of its Rust dependencies and just return the
non-Rust NativeLinkable deps. The Rust deps are compiled as rlibs
and are statically linked into the cdylib, so we just care about
the non-Rust ones.

Test Plan: CI

Reviewed By: andrewjcg

fbshipit-source-id: 60d9e8b