Better exception processing for the parser
Summary:
Buck parser may swallow InterruptedException if it is reported as a cause and not as a first level future interrupt exception. That causes parser to report incorrect exit code (build_error instead of signal_interrupt).
With this change it will check for specific exception type within the whole exception cause chain.
ThrowableCauseIterable is the new iterator implementation with its unit tests that supports this change. This infrastructure will also be used for unified exception processing on top-level in Main.

Test Plan: unit tests, cancel parser job

Reviewed By: ttsugriy

fbshipit-source-id: 4fab223