Changes to address docs issues with build, target, and target pattern
Summary:
- Reorganize Concepts TOC to make **Build Target** and **Build Target
  Patterns** adjacent.
- Make the intro paragraph to **Build Rule** less Android-centric.
- Change HTML hyperlinks to the Closure Template links syntax.
- Fix character formatting: italics (rather than bold) is used to
  introduce new terms such as _Build Rule_ or _Build File_.
- In **Build Target**, remind the reader that build files are _always_
  named `BUCK`.
- Rename and rewrite the section **Command-line Pro Tips**.
- Add See Also from **Build Target** to **Build Target Pattern**.
- Clarify that **buck build** takes a build target pattern. Provide a
  formal syntax block in addition to the already-provided example.
- In **buck build**, rewrite `if...not` construction as `unless`.

- **QUESTION**: We allow build target patterns to specify visibility, but do we allow them to specify dependencies.

Test Plan:
Before sending out my diff, I reviewed the Buck documentation in
rendered form using a local web server. The script for doing this in in
the Buck source tree. Here is the procedure.

Start the web server:

    $ cd ~/local/buck/docs
    $ ./soyweb-local.sh

    Listening on /0:0:0:0:0:0:0:0:9811

Go to your web browser and listen on the port that the server is running
on. In the above example, that would be:

    http://localhost:9811/

On OSX, you can install `pandoc` using `brew`:

    brew install pandoc

I also run `aspell`, which is a command-line spell checker. On OSX, you
can install `aspell` using `brew`:

    brew install aspell

Reviewed By: philipjameson, ttsugriy

fbshipit-source-id: 3fbbe13