Add DEBUG_ISOLATED_IN_PROCESS strategy
Summary:
Adds a new MBR build strategy DEBUG_ISOLATED_IN_PROCESS.

This strategy will construct a separate isolated build directory for each rule. The rule will
be serialized to data files in that directory, and all inputs required (including buck configs)
will be materialized there and then the rule will be deserialized within this process and run
within that directory. The outputs will be copied back to the real build directory.

This requires that we can correctly determine inputs and outputs of a rule and that we can
run it in a context where it doesn't have access to other files or to other BuildRule
instances in memory.

Test Plan: ModernBuildRuleStrategyIntegrationTest covers this new strategy in various scenarios.

Reviewed By: ttsugriy

fbshipit-source-id: 5294266