Do not include cell root path in BuildTargetMacro hashCode
Summary:
This is a temporary workaround for an issue reported in https://github.com/facebook/buck/issues/1800

The solution is to by pass `BuildTarget.hashCode` and use `BuildTarget`'s fully qualified name in `BuildTargetMacro.hashCode`. Also, modified `equals` to be consistent with `hashCode`.

Test Plan: Tried to build using instructions provided in PR, verified that the rule keys of the genrule are the same after the fix (and different before the fix.)

Reviewed By: sbalabanov

fbshipit-source-id: 4470069