account for removed classes in obfuscation translation
Summary:
if a class foo.a is removed and a different class foo.bar is mapped to foo.a, then returning the input will cause both the removed class foo.a and foo.bar to be mapped to foo.a
This adds an optional nullable obfuscation translator the returns null if a class is not in the map.  This relies on the behavior that classes that aren't renamed remain in the map.

Test Plan: tested on local build

Reviewed By: longinoa

fbshipit-source-id: ad7cdc9