Fix file descriptor leak.
Summary:
If directory content stream is not closed explicitly, it can take a
long time for the GC to free those FDs, which can lead to OS running out of
them. This particular instance is not a huge deal, because it only happens on
exception, but still needs to be fixed.

Test Plan: CI

Reviewed By: philipjameson

fbshipit-source-id: 9cc26df