Fix file descriptor leak in tests.
Summary:
If directory content stream is not closed explicitly, it can take a
long time for the GC to free those FDs, which can lead to OS running out of
them. This particular instance is not too bad because the test should be short
lived, but it's a bad example to follow and can still lead to FD exhaustion
when running as part of a large suite of tests.

Test Plan: CI

Reviewed By: philipjameson

fbshipit-source-id: 134dd11