Add BuildRuleResolver to NativeLinkTarget methods
Summary:
Adding `BuildRuleResolver` and related parameters to `NativeLinkTarget` methods. This is a necessary step on the path to changing the `BuildRule`s and related classes to no longer hold on to `BuildRuleResolver`s and related objects in fields.

This is needed for incremental action graph generation, as we have issues with the wrong `BuildRuleResolver` being used in certain cases, as old `BuildRuleResolver`s are being held by helper objects (we already update the `BuildRuleResolver`s that `BuildRule`s directly hold on to). We also need to eventually do this work for build rules with providers anyways.

Test Plan: CI

Reviewed By: bobyangyf

fbshipit-source-id: a275483