`WorkerProcessTool`: Allow to start up workers in parallel
Summary:
`WorkerProcessTool` lazily starts worker processes up to its specified capacity. This was subject to lock contention.

When no free worker is available, the behaviour was
- enter a synchronized block
- check the number of already started processes
- if under capacity, start up a new worker process
- add it to the list of started processes

This serialized starting new processes, as checking the number of processes is blocked until any previously started process becomes available.

In `WorkerProcessPoolFactory`, starting up a new process involves waiting for the handshake.

Here, we only synchronize access to the list of started processes, and push an empty atomic reference.
As soon as the new worker process becomes available, the reference is updated.

This is compatible with the current contract, specifically because
- it is not allowed to close down a pool before all tasks have finished running. I.e., the situation where the pool is shut down before all workers have started is going to fail a precondition, instead of leaving orphaned processes behind.
- starting up more workers than the maximum capacity is avoided by inserting empty atomic references into the list, increasing its length in the locked block.

Test Plan: new tests to guarantee the desired behaviour.

Reviewed By: ttsugriy

fbshipit-source-id: 18a951f