Don't generate ABIs if the build is going to fail anyway
Summary:
If `javac` finds errors in the code, some of its data structures may be left in a bad state.
Accessing those in the course of generating an ABI jar can lead to crashes. The build is going to
fail anyway, so we should jsut skip the work.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: e2c997a