Add retrieval of collection of Providers to BuildRule
Summary:
For AG construction, we want to access the collection of providers for BuildRules.

During AG creation, one BuildRule creation step will could desire arbitrary number of Providers from a dependent BuildRule
The async API will then supply the ProviderCollection of the dependent BuildRule to the BuildRule creation step, so that
it can access any of the Providers it wants in the callback.
Without this, callbacks trying to get dependencies will have to deal with Maps of ProviderKey to Providers, and in case of
retrieving Providers from multiple dependencies, Maps of BuildRuleKey to Maps of ProviderKey to Providers, which isn't very clean.

Test Plan: unit

Reviewed By: jtorkkola

fbshipit-source-id: b808d2e