Ignore ui.superconsole and color.ui for daemon restarts
Summary: If UI flags are changed, we don't need to restart the daemon

Test Plan:
Prior to this version of buck, changing ui.superconsole or color.ui on the command line would result in this line in the logs:
```
[2018-03-15 08:44:05.203][warn ][command:null][tid:208][com.facebook.buck.cli.DaemonLifecycleManager] Shutting down and restarting daemon on config or directory resolver change (Cell{filesystem=com.facebook.buck.io.filesystem.impl.DefaultProjectFilesystem@e9a6e370 (projectRoot=/Users/pjameson/local/buck, hash(blackListedPaths)=-1780814060), buckConfig=com.facebook.buck.config.BuckConfig@ea58b726 (config=com.facebook.buck.util.config.Config@ea58b707)} != Cell{filesystem=com.facebook.buck.io.filesystem.impl.DefaultProjectFilesystem@e9a6e370 (projectRoot=/Users/pjameson/local/buck, hash(blackListedPaths)=-1780814060), buckConfig=com.facebook.buck.config.BuckConfig@9a804763 (config=com.facebook.buck.util.config.Config@9a804744)})
```

With this version, chaning ui.superconsole or color.ui does not result in that message, nor reparsing messages. Changing other arbitrary flags /does/ cause that to appear. Chaning color.ui and ui.superconsole still seems to change the UI on each run, even though the daemon is not restarted.

Reviewed By: ttsugriy

fbshipit-source-id: 7940c3c