Add Vim session-state files to .gitignore for Buck
Summary:
It is possible to configure Vim so that it automatically saves the state
of your editing session when you exit. The state is saved to a file
called

    Session.vim

in the current directory.

This change configures Git to ignore these Vim session files.

Test Plan: I made the proposed `.gitignore` change in a branch and it behaved as expected. That is, git ignored multiple instances of `Session.vim` in the Buck tree, but did not ignore other unstaged files in the tree that were not identified in `.gitignore`.

Reviewed By: ttsugriy

fbshipit-source-id: fa9a91c