Add a way to have platform dependent binaries.
Summary:
In case `apple_bundle` should pack different binaries depending
on the target platform, `platform_binary` can be used to specify which
one should be used for which.

For example,

  apple_bundle(
    name = "foo",
    binary = ":default_binary",
    platform_binary = [
      ("iphoneos-armv7", ":armv7-binary"),
      ("iphoneos-arm64", ":arm64-binary")]
    ),
    ...
  )

The attribute is intentionally not documented for the time being, since
in the long run we'd like to unify handling of platform dispatched deps
using a separate rule (something like `select` from Bazel).

Test Plan: CI

Reviewed By: philipjameson, styurin

fbshipit-source-id: 401dab9