Patch Skylark to report location of unconfigured functions.
Summary:
Previously, if for some reason of the Skylark functions is not configured
before it's used, the error would be something very unhelpful as:
```
java.lang.IllegalStateException: Function struct was not configured
```
and a Java strack trace with original of this exception. But unfortunately it
does not give any context as to where this function was used.

This patch replaces `IllegalStateException` with `EvalException` that carries
location context that is used to create nice error messages pointing at where
in the build or extension file this function was invoked.

Test Plan: CI

Reviewed By: philipjameson

fbshipit-source-id: 78563cb