Make TransitiveCxxPreprocessorInputCache not hold onto BuildResolver
Summary: A step toward not holding `BuildRuleResolver` references in `BuildRule`s and related objects. Extracts a `TransitiveCxxPreprocessorInputCache` class which encapsulate the underlying cache object, so that we can pass in the value loader function during `get`. Previously this used a `LoadingCache`, which took the value loader (which has a `BuildRuleResolver` dependency) in during construction.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: c85c4a8