Add @Before to EndToEndRunner
Summary:
Allows us to run per-class assumptions (for things like upcoming Android tests), kept consistent with how the default JUnit runner does it.

I'll get to adding after, expected exceptions, etc. in future diffs, but this one is needed for android tests.

Test Plan: Ran EndToEndRunnerTest with an before that just instructed the tests to fail. It did that.

Reviewed By: styurin

fbshipit-source-id: 8933ac4