`WorkerProcessPoolTest`: Also accept `WAITING`
Summary:
`WorkerProcessPoolTest` is waiting for a test thread to either terminate, or get blocked. This is achieved by waiting for the thread to either reach the `TERMINATED` or `BLOCKED` state.

Occasionally, the thread will be in a `WAITING` state, and the test will time out after 90 seconds.

Here, we also check whether for `WAITING`, and assert once reached.

Test Plan: Running the test a couple of times did not block. However, we will have to monitor timeouts on CI over a certain amount of time to find out whether this change is enough

Reviewed By: bobyangyf

fbshipit-source-id: e4d26dd