Add TestRule support to the EndToEndRunner
Summary:
Add Rule support for E2ERunner to bring it more in line with default JUnitRunner.

We are not supporting MethodRules as they are deprecated functionally (as mentioned in comment)

After this, we will be in line with the default JUnitRunner minus the Test timeout parameters, which is deprecated, and can be mimicked with TestRules.

Reviewed By: sbalabanov

fbshipit-source-id: 1de8e73