Use compilation-database to view compiler flags
Summary:
Update `buck build` documentation to explain how to use the compilation
database to view the flags that Buck is sending to the C++ compiler.

In the next week or two, I will create a new section in the docs that describes
common tasks--and move this content there. But this is a good place for
it for now.

Reviewed By: ttsugriy

fbshipit-source-id: e98f143