Remove InitializableFromDisk and HasRuntimeDeps from MBR
Summary:
The idea with these is that we could derive runtime deps the same as
inputs/outputs/deps and could implement InitializableFromDisk in some
way too. However, this isn't really implemented yet and for easier
adoption we can allow subclasses to implement those themselves and maybe
migrate to a better system at some future time.

Reviewed By: ttsugriy

fbshipit-source-id: 7630c90