#956 Add kotlin annotation processor (kapt) support.
Summary:
This is still a major work in progress but I wanted to get the ball rolling. Currently this is building but fails at runtime to find the generated classes. I have only tested running it on an Android Project.

A huge shoutout to my former colleague runningcode who helped tremendously on this. #956 outlines the steps that he discovered when running kapt via Gradle.

I've created a simple [tool](https://github.com/cwoodwar6/kapt-cli) to better understand what steps are required when kapt is plugged into the compiler. I'll continue to update this PR and the tool as I make more progress. Feedback is welcome, looking for suggestions early so we can get this added as soon as possible.
Closes https://github.com/facebook/buck/pull/1584

Reviewed By: jkeljo

Pulled By: jkeljo

fbshipit-source-id: 47a9598