Eliminate expensive hash code
Summary:
`DependencyType.putWithMerge` is called at least once (and usually many times) on every `IJModule` in the project. The calling code uses
a `HashMap`, which leads to a lot of CPU time spent calculating the hash code of an `IJModule`. This is expensive due to the large number of
lists of strings contained in the `IJModule`. This change switches to using the module's name (which is guaranteed to be unique because it is based
on the filepath) as the source of truth for `equals` and `hashCode`.

Reviewed By: mkillianey, styurin

fbshipit-source-id: fdffb6d