Remove unnecessary parsing work
Summary:
When creating an intellij project, we need to know the package of our source files. Since not all source files are
located in their canonical package location in the folder hierarchy, we need to parse the source in order to determine
the package. Prior to this diff, the source was parsed in its entirety, but in reality, we only need the `package ...;`.

This diff saves I/O work as well as java AST parsing work by only parsing the first package line.

Reviewed By: styurin

fbshipit-source-id: 8522ebc