Support Apple TV simulator
Summary:
Supporting the Apple TV simulator in a macOS computer is simple by using the support already provided to support iPhone simulators.

A new default Apple TV simulator name is provided to check for in case of the Apple TV Simulator SDK being used. This mirrors the current support for iPhone simulator which checks for a default iPhone simulator name. One can provide a different name with the usual command line switch if needed (to support 4k devices, for example).

Includes tests (and a fixture) to check for the Apple TV simulator basic use cases. The tests are copied from the iPhone simulator ones and fixed them to pass the assumptions and not be skipped (the iPhone simulator ones will be fixed in a different diff).

Reviewed By: ttsugriy

fbshipit-source-id: 75fd891