propagate extra deps for go_test in order to use buildtargets in srcs
Summary:
philipjameson fixed the issue (D6839251) where we couldn't use buildtargets in srcs. For go test the .withoutExtraDeps is being used in test-main-src part of the code. This is because the filtered deps are within declared deps (if that makes sense to you).
I am extending the dependencies to include missing targets

Reviewed By: styurin

fbshipit-source-id: 62b5c29