Make function args optional.
Summary:
Currently args are required even though some of the functions
have no arguments and have to pass the same meaningless `No arguments`
parameter.

Now, callers can just omit args altogether. If we decide to generate some
placeholder to explicitly tell that some function does not have arguments
it's easier to implement it by checking if `$args` is not set and provide
the same template.

Reviewed By: sbalabanov

fbshipit-source-id: a997c7f