Make dx.max_threads parameter to do what the name says
Summary: dx.max_threads parameter is misleading. While the name implies that it should limit the number of threads for in-proc dexer (Android tool that converts Java bytecode to Android-specific bytecode to run on Android platform), in fact it just sets the number of threads. So this change will have max_threads to actually cap the number of threads detected by heuristics. A separate parameter `threads` allows to set a fixed number of threads following the previous behavior of `max_threads`. This is potentially a breaking change, however I believe it will not cause any serious casualties. With this change the number of dexing threads can only decrease and fallback to the value inferred from hardware profile, which is in most cases a good thing for both performance and memory.

Reviewed By: bobyangyf

fbshipit-source-id: a673ba2