go: switch to custom platform definitions via `.buckconfig`
Summary:
This updates the platform handling for Go to be more similar to how other
languages work, where `go#<name>` sections in `.buckconfig` are used to
defined custom platforms.  In this diff, the `GoPlatformFlavorDomain`, which
previously created platforms on the fly named after `<os>_<arch>`, is merged
into `GoPlatformFactory` where platforms are defined exclusively via
`.buckconfig` entries (and named accordingly).

Reviewed By: ttsugriy

fbshipit-source-id: 9b3b855