Minions now split into low and standard spec hardware categories. Re-allocation of nodes always goes to standard spec minions.
Summary:
Minions and work allocator are now aware of 'minion type', passed as a flag to distbuild command, which splits minions into low spec and standard spec hardware categories.

When a minion (of any type) fails, all of the work it had pending is allocated to standard spec minions during future requests.

Reviewed By: ruibm

fbshipit-source-id: c2eed5b