Add BuildRuleKey and BuildRuleContextWithEnvironment for AG construction in AsyncTransformationEngine
Summary:
We will use AsyncTransformationEngine to perform TargetNode to BuildRule transformation.
BuildRuleKey will be used to pass BuildRuleCreationContext and the desired BuildTarget
around to specify what BuildRule to compute, and any necessary information to construct the BuildRule
passed down from the parent. It will also act as the cache key for incrementality.

BuildRuleContextWithEnvironment will be used to wrap the BuildRuleCreationContext for BuildRule
creation functions to access dependencies in a restricted way, and allow BuildRule creation
to trigger asynchronous creations of dependent BuildRules via the engine.

Reviewed By: jtorkkola

fbshipit-source-id: 178b702