Generate manifests for modules if module_manifest_skeleton is specified in android_binary
Summary:
This is the first step in generating resources for module by running aapt on modules. This will generate an android manifest that is suitable for distributing feature APKs.
Next steps will be to allow aapt2 mode, allow resources and assets to be located in the module that depends on them.

fbshipit-source-id: f1ae143