Make Daemon not hold onto BroadcastEventListener and make EventBus dependency explicit
Summary:
Daemon previous holds on to a `BroadcastEventListener`, which is set
and then unset with the proper event bus every time we run buck due to a
dependency from a daemon lifetime `Parser` that requires sending `BuckEvents`.

This is a hidden dependency on the `EventBus`, and isn't nicely implemented.
Instead, we should pass in the current instance of `EventBus` when needed
during a build.

Reviewed By: ttsugriy

fbshipit-source-id: 5c14ba7