Convert Python to a module
Summary:
To convert Python functionality to a module I had to extend the framework to support external resources. The external resources are not packaged inside the module's jar, but placed along with it in `resources` directory.

In case of Python it's used for `make_pex.py` and `pex.pex`. The type of the resource depends on the type of the buck build (buck built by ant uses `make_pex.py` while buck built by buck uses `pex.pex`.) There are no changes in this logic, we still deduct the type of the resource.

`make_pex.py` can also be called when running tests inside IntelliJ. With that change resources are copied to module directory which has a different location than the original place. Since `make_pex.py` adjusts Python path when it's called directly as python file, that logic was extended to cover both cases when it's called from inside of modules directory (by buck built by ant) and from the original location (when called from IntelliJ). Placing directories that cover both cases works fine even though they don't exist, Python just ignores non-existing directories. Alternative approach might be to examine the directories to find the root of the repo or pass an environment variable, but it would either increase complexity of add unnecessary operations.

`DefaultPexToolProvider` was updated to encapsulate knowledge about resource names. The result of these changes is removal of `path_to_pex` from python wrappers.

Reviewed By: ttsugriy

fbshipit-source-id: ab3a3a9