go: don't strongly couple C/C++ and Go platform configurations
Summary:
Recent changes to support multi-platform Go builds used the names of configured
C/C++ platforms to define the set of Go sections to search for Go platform
definitions.  This meant that for a custom Go platform to be defined, there had
to be an existing C/C++ platform with the same name.

This diff decouples Go platform definition from C/C++ platforms, and adds a new
Go section parameter, `cxx_platform`, which is used to tie a given Go platform
to a C/C++ platform.  This alows for more flexible Go platform configuration,
including a many Go platform to one C/C++ platform mapping, which wasn't possible
before.

Reviewed By: bobyangyf

fbshipit-source-id: 7862d12