Make VersionedTool take a PathSourcePath instead of Path
Summary:
Buck doesn't understand how to interpret a Path. It does understand how
to interpret a PathSourcePath. This changes VersionedTool to use the
latter.

This allows things like deriving inputs/deps, doing
serialization/deserialization, correctly resolving paths (because even a
"relative" PSP includes the ProjectFilesystem that it is relative
to).

Reviewed By: ttsugriy

fbshipit-source-id: 88f5778