Make CxxPreprocessAndCompile take an output name instead of path
Summary:
Build rules shouldn't typically allow their output paths to be so freely
configured. There are logical restrictions that Buck requires for
outputs that are much harder for a build rule to enforce when it can be
configured by callers and there ends up being these hidden implicit
assumptions that are built up between these places.

Also, this moves the outputs into a directory that is
unique to the rule instead of being shared with other rules.

Reviewed By: styurin

fbshipit-source-id: 9c21c2c