Add support for Immutables
Summary:
Adds support for Value.Immutable references from ModernBuildRules.

This one is a bit tricky, we still interact only with the fields of the
generated immutable implementation. We use either a constructor or the Builder
to determine which fields are actually important (the immutable implementations
include some extra fields to handle features like Value.Derived and
Value.Lazy fields).

Reviewed By: ttsugriy

fbshipit-source-id: 4cb6378