Allow customization of inputs/deps derivation and serialization
Summary:
This adds two annotations, `CustomClassBehavior` and
`CustomFieldBehavior` and then changes input derivation, deps
derivation and serialization/deserialization to detect these annotations
to allow customization of those phases.

CustomFieldBehavior allows customization of the behavior for just a
specific field. It's common to have a complicated object that could use
the default behavior for most of its fields but needs to handle just a
few in a custom way. Being able to override the behavior for a specific
field makes this less error prone (because the other fields are handled
automatically) and can be simpler.

CustomClassBehavior allows custom behavior specified for the entire
class. For inputs/deps it's easier to just use
HasCustomInputs/DepsLogic, but for deserialization we can't just have
the class implement an interface (because there isn't an instance of the
class at that point). We could just create an instance of the object
using objenesis and then call a function on an uninitialized object, but
that makes it really difficult to have custom deserialization for
Value.Immutable objects. Also, it's really nice how CustomClassBehavior
mirrors the use of CustomFieldBehavior.

Reviewed By: ttsugriy

fbshipit-source-id: 694856b